+++ src/constants/weekDay.tsx
@@ -1,0 +1,13 @@
+type ObjType = {
+  [index: number]: string;
+};
+const WEEKDAY: ObjType = {
+  0: "일요일",
+  1: "월요일",
+  2: "화요일",
+  3: "수요일",
+  4: "목요일",
+  5: "금요일",
+  6: "토요일",
+};
+export default WEEKDAY;
